운영자기능은 모두 로그인된 아이디가 admin이여야 사용할 수 있다는 조건절을 걸었으며, 댓글은 작성자 아이디==로그인아이디

+++ SemiProject/src/board/controller/NoticeDeleteAction.java
@@ -2,59 +2,78 @@ package board.controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import board.model.*;
 import common.controller.AbstractController;
+import member.model.MemberVO;
 
 public class NoticeDeleteAction extends AbstractController {
 
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		
-		// if( "admin".equalsIgnorecase(${sessionScope.loginuser.userid}) ) {}
-		
-		String method = request.getMethod(); // "GET" OR "POST"
-		
-		if("POST".equalsIgnoreCase(method)) { // POST로 들어왔다는 것은 팝업창에서 삭제하겠다고 동의한 것이다.
+		HttpSession session = request.getSession();
+		MemberVO loginuser = (MemberVO) session.getAttribute("loginuser");
+		if( loginuser != null && "admin".equals(loginuser.getUserid())) {
+			String method = request.getMethod(); // "GET" OR "POST"
 			
-			int boardno = Integer.parseInt(request.getParameter("boardno"));
-			
-			InterBoardDAO bdao = new BoardDAO();
-			
-			int n = bdao.deleteNotice(boardno); // 공지사항 게시글 삭제하기
-			
-			String message = "";
-			String loc = "";
-			
-			if(n==1) { // 삭제 성공
-				message = "삭제 성공";
-				loc = request.getContextPath()+"/board/notice.sh";
+			if("POST".equalsIgnoreCase(method)) { // POST로 들어왔다는 것은 팝업창에서 삭제하겠다고 동의한 것이다.
 				
-				request.setAttribute("message", message);
-				request.setAttribute("loc", loc);
-				super.setViewPage("/WEB-INF/deleteMsg.jsp");
+				int boardno = Integer.parseInt(request.getParameter("boardno"));
+				
+				InterBoardDAO bdao = new BoardDAO();
+				
+				int n = bdao.deleteNotice(boardno); // 공지사항 게시글 삭제하기
+				
+				String message = "";
+				String loc = "";
+				
+				if(n==1) { // 삭제 성공
+					message = "삭제 성공";
+					loc = request.getContextPath()+"/board/notice.sh";
+					
+					request.setAttribute("message", message);
+					request.setAttribute("loc", loc);
+					super.setViewPage("/WEB-INF/deleteMsg.jsp");
+				}
+				
+				else { // 삭제 실패
+					message = "삭제 실패!!!!!";
+					loc = request.getContextPath()+"/index.sh";
+					
+					request.setAttribute("message", message);
+					request.setAttribute("loc", loc);
+					super.setViewPage("/WEB-INF/msg.jsp");
+				}
+				
 			}
 			
-			else { // 삭제 실패
-				message = "삭제 실패!!!!!";
-				loc = request.getContextPath()+"/index.sh";
+			else { // GET으로 왔다는 것은 운영자이면서!!, 게시판 상세보기에서 삭제하기를 누른 것이다. 또는 URL로 들어왔거나
 				
-				request.setAttribute("message", message);
-				request.setAttribute("loc", loc);
-				super.setViewPage("/WEB-INF/msg.jsp");
+				String boardno = request.getParameter("boardno");
+				
+				request.setAttribute("boardno", boardno);
+				
+				super.setViewPage("/WEB-INF/board/noticeDelete.jsp");
+				
 			}
+			
 			
 		}
 		
-		else { // GET으로 왔다는 것은 운영자이면서!!, 게시판 상세보기에서 삭제하기를 누른 것이다. 또는 URL로 들어왔거나
+		else {
 			
-			String boardno = request.getParameter("boardno");
+			String message = "잘못된 접근입니다.";
+			String loc = request.getContextPath()+"/index.sh";
 			
-			request.setAttribute("boardno", boardno);
+			request.setAttribute("message", message);
+			request.setAttribute("loc", loc);
 			
-			super.setViewPage("/WEB-INF/board/noticeDelete.jsp");
+			super.setViewPage("/WEB-INF/msg.jsp");
 			
 		}
+
 		
 		////////////////////////////////////////////////////////////////////////////
